Add files via upload
In this project, We were required to develop this system with the help of python programming language and the Integrated Development and Learning Environment (IDLE). We were required to develop this program with the help of an algorithm and flowchart. In this system, we were required to develop a properly functioning system to view, rent, and return the costume. This system is very useful to the owners of Rental shops where people borrow and return items. This program can help those owners by keeping records of borrowing, returns, and transactions. Instead of keeping records on paper, this system can be very helpful to rental shop owners.   
 
The goals and objectives of this project are very clear. This is the system for the Rental shop management system. People can view the available costumes, select their favorite costumes as per their desire, and can rent them for a certain time and return the costume in proper time as people do in real life in the rental shops. This system also helps in keeping the record of every transaction.  This system eradicates the problem of keeping every record on physical paper and keeping it safe as its loss can disturb the entire business. User can select their desirable costume from the list and enter their name in order to rent the costume. Then a bill is generated for that costume with the customer name, customer Id, Name of the costume, Brand of the costume, Price of the costume, the total quantity of the costume, and the total price of the costume. Meanwhile returning the costume, the system automatically finds the bill which was generated while renting the costume with the user name.